x86: Address sparse warnings about missing UL tags for constants
Automatic type conversion saved us in all these cases so far, but better
avoid surprises in the future and another finding turned out to be a bug
in fact.

JAILHOUSE_BASE requires special wrapping as it is also used in assembly
(the linker script) where the UL tag is not understood.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>